chore(menu-builder): convert to es6
